backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
nanoid==2.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.1.1
djangorestframework==3.15.2
PyJWT==2.10.1
django-cors-headers==4.7.0